Feat: Remove physx specific deformable schemas
