+++ scripts/tutorials/01_assets/run_deformable_object.py
@@ -40,7 +40,7 @@ import torch
 import warp as wp
 
 # deformables supported in PhysX
-from isaaclab_physx.sim import DeformableBodyMaterialCfg, DeformableBodyPropertiesCfg
+from isaaclab_physx.sim import DeformableBodyMaterialCfg, SurfaceDeformableBodyMaterialCfg
 
 import isaaclab.sim as sim_utils
 import isaaclab.utils.math as math_utils
@@ -70,24 +70,25 @@ def design_scene():
     # 3D Deformable Object
     cfg = DeformableObjectCfg(
         prim_path="/World/env_.*/Cube",
-        # spawn=sim_utils.MeshCuboidCfg(
-        #     size=(0.2, 0.2, 0.2),
-        #     deformable_props=sim_utils.DeformableBodyPropertiesCfg(rest_offset=0.0, contact_offset=0.001),
-        #     visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.5, 0.1, 0.0)),
-        #     physics_material=sim_utils.DeformableBodyMaterialCfg(poissons_ratio=0.4, youngs_modulus=1e5),
-        # ),
-        spawn=TetMeshCuboidCfg(
+        spawn=sim_utils.MeshCuboidCfg(
             size=(0.2, 0.2, 0.2),
-            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.2, 0.8, 0.2)),
+            deformable_props=sim_utils.DeformableBodyPropertiesCfg(rest_offset=0.0, contact_offset=0.001),
+            visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.5, 0.1, 0.0)),
+            physics_material=DeformableBodyMaterialCfg(poissons_ratio=0.4, youngs_modulus=1e5),
+            # physics_material=SurfaceDeformableBodyMaterialCfg(poissons_ratio=0.4, youngs_modulus=1e4, surface_thickness=0.001, surface_bend_stiffness=1e0, surface_shear_stiffness=1e0, surface_stretch_stiffness=1e0),
         ),
+        # spawn=TetMeshCuboidCfg(
+        #     size=(0.2, 0.2, 0.2),
+        #     visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.2, 0.8, 0.2)),
+        # ),
         init_state=DeformableObjectCfg.InitialStateCfg(pos=(0.0, 0.0, 1.0)),
-        density=500.0,
-        tri_ke=1e5,
-        tri_ka=1e5,
-        tri_kd=1e-4,
-        edge_ke=100.0,
-        edge_kd=1e-2,
-        particle_radius=0.005,
+        # density=500.0,
+        # tri_ke=1e5,
+        # tri_ka=1e5,
+        # tri_kd=1e-4,
+        # edge_ke=100.0,
+        # edge_kd=1e-2,
+        # particle_radius=0.005,
     )
 
     cube_object = DeformableObject(cfg=cfg)
@@ -110,7 +111,7 @@ def run_simulator(sim: sim_utils.SimulationContext, entities: dict, origins: tor
     count = 0
 
     # Nodal kinematic targets of the deformable bodies
-    nodal_kinematic_target = wp.to_torch(cube_object.data.nodal_kinematic_target).clone()
+    # nodal_kinematic_target = wp.to_torch(cube_object.data.nodal_kinematic_target).clone()
 
     # Simulate physics
     # for _ in range(200):
@@ -134,9 +135,9 @@ def run_simulator(sim: sim_utils.SimulationContext, entities: dict, origins: tor
             cube_object.write_nodal_state_to_sim_index(nodal_state)
 
             # Write the nodal state to the kinematic target and free all vertices
-            nodal_kinematic_target[..., :3] = nodal_state[..., :3]
-            nodal_kinematic_target[..., 3] = 1.0
-            cube_object.write_nodal_kinematic_target_to_sim_index(nodal_kinematic_target)
+            # nodal_kinematic_target[..., :3] = nodal_state[..., :3]
+            # nodal_kinematic_target[..., 3] = 1.0
+            # cube_object.write_nodal_kinematic_target_to_sim_index(nodal_kinematic_target)
 
             print("----------------------------------------")
             print("[INFO]: Resetting object state...")
@@ -144,12 +145,12 @@ def run_simulator(sim: sim_utils.SimulationContext, entities: dict, origins: tor
         # update the kinematic target for cubes at index 0 and 3
         kinematic_cubes = [0, 3]
         # we slightly move the cube in the z-direction by picking the vertex at index 0
-        nodal_kinematic_target[kinematic_cubes, 0, 2] += 0.2 * sim_dt
+        # nodal_kinematic_target[kinematic_cubes, 0, 2] += 0.2 * sim_dt
         # set vertex at index 0 to be kinematically constrained
         # 0: constrained, 1: free
-        nodal_kinematic_target[kinematic_cubes, 0, 3] = 0.0
+        # nodal_kinematic_target[kinematic_cubes, 0, 3] = 0.0
         # write kinematic target to simulation
-        cube_object.write_nodal_kinematic_target_to_sim_index(nodal_kinematic_target)
+        # cube_object.write_nodal_kinematic_target_to_sim_index(nodal_kinematic_target)
 
         # write internal data to simulation
         cube_object.write_data_to_sim()
