+++ source/isaaclab/isaaclab/sim/__init__.pyi
@@ -22,11 +22,13 @@ __all__ = [
     "activate_contact_sensors",
     "define_articulation_root_properties",
     "define_collision_properties",
+    "define_deformable_body_properties",
     "define_mass_properties",
     "define_mesh_collision_properties",
     "define_rigid_body_properties",
     "modify_articulation_root_properties",
     "modify_collision_properties",
+    "modify_deformable_body_properties",
     "modify_fixed_tendon_properties",
     "modify_joint_drive_properties",
     "modify_mass_properties",
@@ -39,6 +41,7 @@ __all__ = [
     "CollisionPropertiesCfg",
     "ConvexDecompositionPropertiesCfg",
     "ConvexHullPropertiesCfg",
+    "DeformableBodyPropertiesCfg",
     "FixedTendonPropertiesCfg",
     "JointDrivePropertiesCfg",
     "MassPropertiesCfg",
@@ -182,11 +185,13 @@ from .schemas import (
     activate_contact_sensors,
     define_articulation_root_properties,
     define_collision_properties,
+    define_deformable_body_properties,
     define_mass_properties,
     define_mesh_collision_properties,
     define_rigid_body_properties,
     modify_articulation_root_properties,
     modify_collision_properties,
+    modify_deformable_body_properties,
     modify_fixed_tendon_properties,
     modify_joint_drive_properties,
     modify_mass_properties,
@@ -199,6 +204,7 @@ from .schemas import (
     CollisionPropertiesCfg,
     ConvexDecompositionPropertiesCfg,
     ConvexHullPropertiesCfg,
+    DeformableBodyPropertiesCfg,
     FixedTendonPropertiesCfg,
     JointDrivePropertiesCfg,
     MassPropertiesCfg,
