+++ source/isaaclab/isaaclab/sim/spawners/from_files/from_files.py
@@ -12,7 +12,6 @@ import tempfile
 from typing import TYPE_CHECKING
 
 # deformables only supported on PhysX backend
-from isaaclab_physx.sim import schemas as schemas_physx
 from isaaclab_physx.sim.spawners.materials import SurfaceDeformableBodyMaterialCfg
 
 from pxr import Gf, Sdf, Usd, UsdGeom
@@ -372,9 +371,9 @@ def _spawn_from_usd_file(
         prim = stage.GetPrimAtPath(prim_path)
         deformable_type = "surface" if isinstance(cfg.physics_material, SurfaceDeformableBodyMaterialCfg) else "volume"
         if "OmniPhysicsDeformableBodyAPI" in prim.GetAppliedSchemas():
-            schemas_physx.modify_deformable_body_properties(prim_path, cfg.deformable_props, stage)
+            schemas.modify_deformable_body_properties(prim_path, cfg.deformable_props, stage)
         else:
-            schemas_physx.define_deformable_body_properties(prim_path, cfg.deformable_props, stage, deformable_type)
+            schemas.define_deformable_body_properties(prim_path, cfg.deformable_props, stage, deformable_type)
         if cfg.mass_props is not None:
             raise ValueError(
                 """MassPropertiesCfg are not supported for deformable bodies
