--- source/isaaclab_physx/isaaclab_physx/sim/schemas/__init__.py
@@ -1,10 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""Sub-module containing utilities for schemas used in Omniverse for PhysX backend."""
-
-from isaaclab.utils.module import lazy_export
-
-lazy_export()
