+++ source/isaaclab_physx/setup.py
@@ -53,7 +53,6 @@ setup(
         "isaaclab_physx.sensors.frame_transformer",
         "isaaclab_physx.sensors.imu",
         "isaaclab_physx.sim",
-        "isaaclab_physx.sim.schemas",
         "isaaclab_physx.sim.spawners",
         "isaaclab_physx.sim.spawners.materials",
         "isaaclab_physx.test",
